fix(wi-4): merge cross-host forwarder tests into test_server.py
Clear the ``PS-204 §2 orphan-test-file`` violation CI flagged on
PR #132: ``tests/.../test_cross_host_forward.py`` had no matching
src module ``_listen/cross_host_forward.py`` because there is no
such module — the forwarder helper (``_forward_to_remote``) and
the route that invokes it (``node_message_send``) both live in
``_listen/server.py``. The canonical mirror is therefore
``tests/.../test_server.py``.

Per the lead's option (a): merge the tests into the existing mirror
test file rather than create a synthetic src module. Same pattern as
the PR #124 ``test_no_channel_group`` fix.

Changes:
  * Append the cross-host loopback section + ``cross_host_env``
    fixture (renamed from ``isolated_env`` to avoid the existing
    fixture's name; broader scope so the rename names the actual
    difference) + ``missing_peer_token_response`` fixture + the
    three loud-502 single-assert tests at the bottom of
    ``test_server.py``.
  * Lift ``SHARED_TOKEN`` constant + ``write_peer_token`` /
    ``state_db`` / ``state_db_nodes.record_lineage`` imports to
    the top of the file so both halves share them.
  * Delete the orphan ``test_cross_host_forward.py``.

Net diff: -423 / +396 lines.

Local verification (``/opt/venv-agent``):
  * ``pytest test_server.py`` → 48 passed (was 41 in the original
    file; +5 from the WI-4 section — 2 forwarder + 3 loud-502).
  * ``scitex-dev ecosystem audit-python-apis`` → ``SUCC: no
    Python API violations``.
  * ``scitex-dev ecosystem audit-project`` no longer flags
    ``test_cross_host_forward.py``. The remaining 12 PS-204
    orphans + PS-302 + PS-122 errors are pre-existing baseline on
    ``develop`` (verified earlier in this session by running the
    same audit on the develop tip).
  * ``scitex-dev linter check-files test_server.py`` → 0 errors,
    2 warnings (``9999`` + ``0`` port literals — explicit
    carve-out per the STX-NL001 rule docs: "ports may stay bare").
