+++ tests/scitex_agent_container/_listen/test_server.py
@@ -5,22 +5,49 @@ with real fixture files under ``tmp_path`` and real bearer-token auth.
 Module-level path roots (registry dir, runtime dir, ``$HOME``) are
 re-pointed at the per-test tmp directory by assigning to the module
 attributes — that's configuration, not mocking.
+
+The WI-4 cross-host forwarder tests live at the bottom of this
+file. The forwarder helper (``_forward_to_remote``) and the
+``node_message_send`` route that invokes it both live in
+``_listen/server.py``, so this is the canonical mirror per the
+PS-204 orphan-test-file rule. The cross-host section uses its own
+``cross_host_env`` fixture (broader scope: state.db + peer-tokens
+registry) and a distinct ``SHARED_TOKEN`` constant so the
+WI-4 loopback tests do not collide with the existing route tests'
+``isolated_env`` / ``TOKEN``.
 """
 
 from __future__ import annotations
 
+import asyncio
+import contextlib
+import json
 import os
+import socket
+import threading
 from pathlib import Path
 
+import httpx
 import pytest
+import uvicorn
 import yaml
 from starlette.testclient import TestClient
 
+from scitex_agent_container._listen.peer_tokens import write_peer_token
 from scitex_agent_container._listen.server import create_app
 from scitex_agent_container._runners import _session_state as _ss
 from scitex_agent_container._state import registry as _reg
+from scitex_agent_container._state import state_db
+from scitex_agent_container._state.state_db_nodes import record_lineage
 
 TOKEN = "test-token-abc123"
+
+# WI-4 cross-host forwarder: both apps in the loopback tests run
+# with the same listen token — Q4(b) per-host bearer registry: the
+# forwarder pulls ``peer-tokens/host-a.token`` (= this value) when
+# forwarding to host A. Real deployments mint independent per-host
+# tokens.
+SHARED_TOKEN = "test-token-wi4"
 
 
 # --- Fixtures --------------------------------------------------------------
@@ -559,3 +586,372 @@ class TestUnknownRoute:
         resp = client.get(url, headers=auth_headers)
         # Assert
         assert resp.status_code == 404
+
+
+# ---------------------------------------------------------------------------
+# WI-4 — cross-host forwarder on ``sac listen`` (handoff §4).
+#
+# Per HANDOFF_AGENT_COMMS_2026-05-19.md §4 (WI-4):
+#
+#   Acceptance: a node on host B sends to one on host A; the event
+#   arrives, ACL-checked at A.
+#
+# The end-to-end tests drive two real ``uvicorn`` instances on
+# loopback ports to simulate the two-host topology:
+#
+#   * "host A" — owns the target. ACL is gated here.
+#   * "host B" — the forwarding entry point. Records the target's
+#     instance under ``host="host-a"`` so the resolver routes there.
+#
+# A POST to host B's ``message:send`` for that target arrives on
+# host A's broker. The Q4(b) per-host bearer registry is exercised
+# by the missing-peer-token loud-502 tests at the bottom.
+#
+# No mocks (handoff §0): real SQLite + real ``uvicorn``.
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def cross_host_env(tmp_path: Path):
+    """Isolated state.db + runtime/registry roots + peer-token for host A.
+
+    Broader scope than ``isolated_env`` above (which only redirects
+    HOME / registry / runtime): this fixture also wires up state.db
+    and seeds the WI-4 Q4(b) per-host bearer registry under
+    ``~/.scitex/agent-container/peer-tokens/`` so the forwarder can
+    authenticate at host A.
+    """
+    # Arrange
+    db = tmp_path / "state.db"
+    saved_db_env = os.environ.get("SCITEX_AGENT_CONTAINER_STATE_DB")
+    saved_db_const = state_db.DEFAULT_DB_PATH
+    saved_home = os.environ.get("HOME")
+    saved_reg_const = _reg.REGISTRY_DIR
+    saved_state_const = _ss.DEFAULT_STATE_ROOT
+
+    os.environ["SCITEX_AGENT_CONTAINER_STATE_DB"] = str(db)
+    os.environ["HOME"] = str(tmp_path)
+    state_db.DEFAULT_DB_PATH = db
+    _reg.REGISTRY_DIR = tmp_path / "registry"
+    _ss.DEFAULT_STATE_ROOT = tmp_path / "runtime"
+    state_db.init_schema(db)
+    # WI-4 Q4(b): seed the per-host bearer registry. The forwarder
+    # pulls ``peer-tokens/host-a.token`` to authenticate when
+    # forwarding to host A. ``$HOME`` is already tmp_path so the
+    # default ``~/.scitex/agent-container/peer-tokens/`` lands here.
+    write_peer_token(peer_host="host-a", token=SHARED_TOKEN)
+    try:
+        yield {"db": db, "tmp": tmp_path}
+    finally:
+        state_db.DEFAULT_DB_PATH = saved_db_const
+        _reg.REGISTRY_DIR = saved_reg_const
+        _ss.DEFAULT_STATE_ROOT = saved_state_const
+        if saved_db_env is None:
+            os.environ.pop("SCITEX_AGENT_CONTAINER_STATE_DB", None)
+        else:
+            os.environ["SCITEX_AGENT_CONTAINER_STATE_DB"] = saved_db_env
+        if saved_home is None:
+            os.environ.pop("HOME", None)
+        else:
+            os.environ["HOME"] = saved_home
+
+
+def _free_port() -> int:
+    with contextlib.closing(socket.socket()) as s:
+        s.bind(("127.0.0.1", 0))
+        return s.getsockname()[1]
+
+
+@contextlib.contextmanager
+def _run_loopback(app, port: int):
+    """Spin up uvicorn on a loopback port. The app's
+    ``local_host`` identity is configured at ``create_app`` time
+    (see :func:`scitex_agent_container._listen.server.create_app`).
+    """
+    config = uvicorn.Config(
+        app, host="127.0.0.1", port=port, log_level="warning", ws="none"
+    )
+    server = uvicorn.Server(config)
+    t = threading.Thread(target=server.run, daemon=True)
+    t.start()
+    import time as _time
+
+    deadline = _time.monotonic() + 5.0
+    while not server.started:
+        if _time.monotonic() > deadline:
+            raise RuntimeError("uvicorn loopback did not start in 5s")
+        _time.sleep(0.05)
+    try:
+        yield port
+    finally:
+        server.should_exit = True
+        t.join(timeout=5.0)
+
+
+def _send_payload(text: str, *, from_agent: str) -> dict:
+    return {
+        "jsonrpc": "2.0",
+        "id": "1",
+        "method": "SendMessage",
+        "params": {
+            "message": {
+                "message_id": "m1",
+                "role": "ROLE_USER",
+                "parts": [{"text": text}],
+            },
+            "metadata": {"from_agent": from_agent},
+        },
+    }
+
+
+def test_cross_host_send_forwards_to_target_host(cross_host_env) -> None:
+    """End-to-end: a POST to host B's ``message:send`` for a target
+    pinned to host A arrives on host A's broker.
+    """
+    # Arrange
+    db = cross_host_env["db"]
+    # Register the target as a live instance on host-a.
+    state_db.record_instance_start(
+        name="alice", host="host-a", a2a_port=0, db_path=db
+    )
+    # Permitted-peer is registered as a child of root, so is alice;
+    # they share a group and ACL allows the send.
+    record_lineage(child="permitted-peer", parent="root", db_path=db)
+    record_lineage(child="alice", parent="root", db_path=db)
+
+    host_a_port = _free_port()
+    host_b_port = _free_port()
+
+    # Bind the actual port for host A onto the instances row so the
+    # resolver routes to the right loopback.
+    with state_db.open_db(db) as conn:
+        conn.execute(
+            "UPDATE instances SET a2a_port = ? WHERE name = 'alice'",
+            (host_a_port,),
+        )
+
+    app_a = create_app(token=SHARED_TOKEN, local_host="host-a")
+    app_b = create_app(token=SHARED_TOKEN, local_host="host-b")
+
+    async def driver() -> dict:
+        with _run_loopback(app_a, host_a_port):
+            # Subscribe on host A as alice.
+            ready = asyncio.Event()
+            captured: dict = {}
+
+            async def consume():
+                async with httpx.AsyncClient(timeout=5.0) as ac:
+                    async with ac.stream(
+                        "GET",
+                        f"http://127.0.0.1:{host_a_port}/agents/alice/inbox/stream",
+                        headers={"authorization": f"Bearer {SHARED_TOKEN}"},
+                    ) as sse:
+                        async for line in sse.aiter_lines():
+                            if line.startswith(":"):
+                                ready.set()
+                                continue
+                            if line.startswith("data:"):
+                                captured["event"] = json.loads(
+                                    line[len("data:") :].lstrip()
+                                )
+                                return
+
+            sub = asyncio.create_task(consume())
+            try:
+                await asyncio.wait_for(ready.wait(), timeout=5.0)
+                # Now stand up host B and post to it. WI-4 forwarder
+                # on host B should resolve alice→host-a and forward.
+                with _run_loopback(app_b, host_b_port):
+                    async with httpx.AsyncClient(timeout=5.0) as ac:
+                        resp = await ac.post(
+                            f"http://127.0.0.1:{host_b_port}/agents/alice/message:send",
+                            json=_send_payload(
+                                "hi from b", from_agent="permitted-peer"
+                            ),
+                            headers={
+                                "authorization": f"Bearer {SHARED_TOKEN}"
+                            },
+                        )
+                if resp.status_code >= 400:
+                    raise RuntimeError(
+                        f"forward returned {resp.status_code}: {resp.text!r}"
+                    )
+                await asyncio.wait_for(sub, timeout=5.0)
+            finally:
+                if not sub.done():
+                    sub.cancel()
+                    with contextlib.suppress(BaseException):
+                        await sub
+            return captured.get("event", {})
+
+    # Act
+    event = asyncio.run(driver())
+    # Assert
+    assert event.get("content") == "hi from b"
+
+
+def test_cross_host_forward_preserves_from_agent_metadata(cross_host_env) -> None:
+    """The forwarded event keeps the original ``from_agent`` so
+    host A's ACL can gate on the real sender, not the forwarding
+    host's identity.
+    """
+    # Arrange
+    db = cross_host_env["db"]
+    state_db.record_instance_start(
+        name="alice", host="host-a", a2a_port=0, db_path=db
+    )
+    record_lineage(child="permitted-peer", parent="root", db_path=db)
+    record_lineage(child="alice", parent="root", db_path=db)
+    host_a_port = _free_port()
+    host_b_port = _free_port()
+    with state_db.open_db(db) as conn:
+        conn.execute(
+            "UPDATE instances SET a2a_port = ? WHERE name = 'alice'",
+            (host_a_port,),
+        )
+    app_a = create_app(token=SHARED_TOKEN, local_host="host-a")
+    app_b = create_app(token=SHARED_TOKEN, local_host="host-b")
+
+    async def driver() -> dict:
+        with _run_loopback(app_a, host_a_port):
+            ready = asyncio.Event()
+            captured: dict = {}
+
+            async def consume():
+                async with httpx.AsyncClient(timeout=5.0) as ac:
+                    async with ac.stream(
+                        "GET",
+                        f"http://127.0.0.1:{host_a_port}/agents/alice/inbox/stream",
+                        headers={"authorization": f"Bearer {SHARED_TOKEN}"},
+                    ) as sse:
+                        async for line in sse.aiter_lines():
+                            if line.startswith(":"):
+                                ready.set()
+                                continue
+                            if line.startswith("data:"):
+                                captured["event"] = json.loads(
+                                    line[len("data:") :].lstrip()
+                                )
+                                return
+
+            sub = asyncio.create_task(consume())
+            try:
+                await asyncio.wait_for(ready.wait(), timeout=5.0)
+                with _run_loopback(app_b, host_b_port):
+                    async with httpx.AsyncClient(timeout=5.0) as ac:
+                        await ac.post(
+                            f"http://127.0.0.1:{host_b_port}/agents/alice/message:send",
+                            json=_send_payload(
+                                "x", from_agent="permitted-peer"
+                            ),
+                            headers={
+                                "authorization": f"Bearer {SHARED_TOKEN}"
+                            },
+                        )
+                await asyncio.wait_for(sub, timeout=5.0)
+            finally:
+                if not sub.done():
+                    sub.cancel()
+                    with contextlib.suppress(BaseException):
+                        await sub
+            return captured.get("event", {})
+
+    # Act
+    event = asyncio.run(driver())
+    # Assert
+    assert event.get("from_agent") == "permitted-peer"
+
+
+# ---------------------------------------------------------------------------
+# WI-4 Q4(b): missing peer-token → loud 502 (handoff §0 — no silent drop).
+# ---------------------------------------------------------------------------
+
+
+@pytest.fixture
+def missing_peer_token_response(tmp_path: Path):
+    """Drive a forwarder POST with NO peer-token written for the
+    destination host, so the forwarder must fall through to the loud
+    502 path. Yielded value is the live ``httpx.Response`` so each
+    test can assert one aspect of the failure shape.
+    """
+    # Arrange — fresh tmp env, NO peer-token for host-z.
+    saved_db_env = os.environ.get("SCITEX_AGENT_CONTAINER_STATE_DB")
+    saved_db_const = state_db.DEFAULT_DB_PATH
+    saved_home = os.environ.get("HOME")
+    saved_reg_const = _reg.REGISTRY_DIR
+    saved_state_const = _ss.DEFAULT_STATE_ROOT
+    db = tmp_path / "state.db"
+    os.environ["SCITEX_AGENT_CONTAINER_STATE_DB"] = str(db)
+    os.environ["HOME"] = str(tmp_path)
+    state_db.DEFAULT_DB_PATH = db
+    _reg.REGISTRY_DIR = tmp_path / "registry"
+    _ss.DEFAULT_STATE_ROOT = tmp_path / "runtime"
+    state_db.init_schema(db)
+    record_lineage(child="permitted-peer", parent="root", db_path=db)
+    record_lineage(child="alice", parent="root", db_path=db)
+    state_db.record_instance_start(
+        name="alice", host="host-z", a2a_port=9999, db_path=db
+    )
+    app_local = create_app(token=SHARED_TOKEN, local_host="host-b")
+
+    try:
+        with TestClient(app_local) as client:
+            r = client.post(
+                "/agents/alice/message:send",
+                json=_send_payload("hi", from_agent="permitted-peer"),
+                headers={"authorization": f"Bearer {SHARED_TOKEN}"},
+            )
+        yield r
+    finally:
+        state_db.DEFAULT_DB_PATH = saved_db_const
+        _reg.REGISTRY_DIR = saved_reg_const
+        _ss.DEFAULT_STATE_ROOT = saved_state_const
+        if saved_db_env is None:
+            os.environ.pop("SCITEX_AGENT_CONTAINER_STATE_DB", None)
+        else:
+            os.environ["SCITEX_AGENT_CONTAINER_STATE_DB"] = saved_db_env
+        if saved_home is None:
+            os.environ.pop("HOME", None)
+        else:
+            os.environ["HOME"] = saved_home
+
+
+def test_cross_host_forward_502_when_peer_token_missing(
+    missing_peer_token_response,
+) -> None:
+    """When forwarding to a host whose bearer is NOT in
+    ``peer-tokens/``, the forwarder fails loudly with 502
+    (handoff §0 — no silent drop).
+    """
+    # Arrange
+    r = missing_peer_token_response
+    # Act
+    status = r.status_code
+    # Assert
+    assert status == 502, r.text
+
+
+def test_cross_host_forward_502_body_names_missing_host(
+    missing_peer_token_response,
+) -> None:
+    """The 502 body names the specific peer host so the operator
+    sees which ``sac host add-peer`` to run."""
+    # Arrange
+    r = missing_peer_token_response
+    # Act
+    err = r.json().get("error", "")
+    # Assert
+    assert "host-z" in err
+
+
+def test_cross_host_forward_502_body_carries_add_peer_fix(
+    missing_peer_token_response,
+) -> None:
+    """The 502 body advertises the ``sac host add-peer`` remediation
+    so the loud failure points at the fix."""
+    # Arrange
+    r = missing_peer_token_response
+    # Act
+    err = r.json().get("error", "")
+    # Assert
+    assert "sac host add-peer" in err
